moove some utils code into an util object and refactor some distance classes
