+++ scala/src/main/scala/smile/clustering/distance/Binary/MeanMahanttan.scala
@@ -8,9 +8,8 @@ import scala.math.pow
  **/
 class MeanMahanttan extends BinaryDistance {
 	
-	override def distance(vector1: Array[Int], vector2: Array[Int]) : Double =
-	{
-		val (a,b,c,d) = contingencyTable(vector1, vector2)
+	override def distance(vector1: Array[Int], vector2: Array[Int]) : Double = {
+		val (a,b,c,d) = BinaryUtils.contingencyTable(vector1, vector2)
 		(b + c).toDouble / (a + b + c + d)
 	}
