+++ scala/src/main/scala/smile/clustering/distance/Binary/ShapeDifference.scala
@@ -8,7 +8,8 @@ import scala.math.pow
 class ShapeDifference extends BinaryDistance {
 	
 	override def distance(vector1: Array[Int], vector2: Array[Int]) : Double = {
-		val (a, b, c, d) = contingencyTable(vector1, vector2)
+		val (a, b, c, d) = BinaryUtils.contingencyTable(vector1, vector2)
 		((a + b + c + d) * (b + c) - pow(b - c, 2)) / pow(a + b + c + d,2)
 	}	
+
 }
