+++ scala/src/main/scala/smile/clustering/distance/Binary/Vari.scala
@@ -7,8 +7,10 @@ import scala.math.pow
  *
  **/
 class Vari extends BinaryDistance {
+
 	override def distance(vector1: Array[Int], vector2: Array[Int]) : Double = {
-		val (a,b,c,d) = contingencyTable(vector1, vector2)
+		val (a,b,c,d) = BinaryUtils.contingencyTable(vector1, vector2)
 		(b + c).toDouble / (4 * (a + b + c + d))
 	}
+	
 }
